% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMPC.best.prevalence.R
\name{DMPC.best.prevalence}
\alias{DMPC.best.prevalence}
\title{Best Prevalence}
\usage{
DMPC.best.prevalence(prev.list, variable, method.dist = "bray")
}
\arguments{
\item{prev.list}{List phyloseq objects with the calculated prevalences for each cutoff}

\item{variable}{Variable to run the model}

\item{method.dist}{Distance/dissimilarity. Default is Bray-Curtis. Suported mesurements from vegan::vegdist()}
}
\description{
Runs Permanova analysis on Bray-Curtis dissimilarity and performes a classification of variables with RandomForests for each prevalence cutoff returned by DMPC.prevalence()
}
\examples{
phylist=DMPC.split.by.variable(restroom, "Environment")
prev=DMPC.prevalence(phylist)
DMPC.best.prevalence(prev, "Environment", method.dist="bray")
}
\keyword{OOB}
\keyword{permanova}
\keyword{prevalence}
