% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMPC.FDR.R
\name{DMPC.FDR}
\alias{DMPC.FDR}
\title{FDR}
\usage{
DMPC.FDR(physeq, pseudo.var, bootstrap, method.dist = "bray")
}
\arguments{
\item{physeq}{Original phyloseq object used for determining the best prevalence cutoff.}

\item{pseudo.var}{Variable to run the model}

\item{bootstrap}{Number of randomizations}

\item{method.dist}{Distance/dissimilarity. Default is Bray-Curtis. Suported mesurements from vegan::vegdist()}
}
\description{
Testing the likelihood of introducing bias by calculations of different prevalence cutoffs.
Randomizes (bootstrap) variable distribuition, performs prevalences cutoff and perMANOVA.
}
\examples{
phylist=DMPC.split.by.variable(restroom, "Environment")
prev=DMPC.prevalence(phylist)
DMPC.best.prevalence(prev, "Environment")
DMPC.FDR(restroom, "Environment", bootstrap = 100, method.dist = "bray")

}
\keyword{FDR}
\keyword{permanova}
\keyword{prevalence}
