% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMPC.OOB.error.R
\name{DMPC.OOB.error}
\alias{DMPC.OOB.error}
\title{Random Forest with full dataset}
\usage{
DMPC.OOB.error(physeq, variable)
}
\arguments{
\item{physeq}{full object phyloseq}

\item{variable}{"variable to run the classification"}
}
\description{
This fucntion runs randonForests on full dataset. If OOB error >0 then proceed with folowing analysis.
}
\examples{
DMPC.OOB.error(restroom, "Environment")
}
\keyword{OOB}
