% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMPC.prevalence.R
\name{DMPC.prevalence}
\alias{DMPC.prevalence}
\title{Prevalences cutoffs}
\usage{
DMPC.prevalence(phylist)
}
\arguments{
\item{phylist}{list of phyloseq objects (one for each variable)}
}
\description{
First calculates which is the highest prevalence possible (no empty OTU table), calculates prevalence for each cutoff by increments of 5%.
After prevalence calculation, merging of phyloseq objects for each variable and cutoff is done when no sample is lost.
When any sample_sums==0 FALSE.
}
\examples{
phylist=DMPC.split.by.variable(restroom, "Environment")
prev=DMPC.prevalence(phylist)
}
\keyword{prevalence}
