% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMPC-package.r
\docType{package}
\name{DMPC}
\alias{DMPC}
\alias{DMPC-package}
\title{Determining Microbial Prevalence Cutoff - DMPC}
\description{
This package identifies the best cutoff for microbial prevalence. It splits phyloseq objects based on variables,
computes prevalences based
Best prevalence cutoff based on Out of Bag error from RandomForests.
}
