% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMPC.split.by.variable.R
\name{DMPC.split.by.variable}
\alias{DMPC.split.by.variable}
\title{Split phyloseq}
\usage{
DMPC.split.by.variable(physeq, variable, drop_zeroes = T)
}
\arguments{
\item{physeq}{object phyloseq}

\item{variable}{"variable to split"}

\item{drop_zeroes}{Remove OTUs with zeroes, defaults to TRUE}
}
\description{
phyloseq objects were sliced by code extracted from metagMisc: v.0.0.4 - DOI 10.5281/zenodo.172500
}
\examples{
phylist=DMPC.split.by.variable(restroom, "Environment")
}
